apps/api/requirements.txt
alembic==1.17.1
email-validator==2.3.0
fastapi==0.115.4
httpx==0.27.1
pydantic==2.10.1
pydantic-settings==2.6.1
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
SQLAlchemy==2.0.44
psycopg[binary]==3.2.12
passlib[bcrypt]==1.7.4
PyYAML==6.0.3
uvicorn==0.32.1
